qcart/requirements.txt
appdirs==1.4.4
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
frozendict==2.3.8
greenlet==2.0.2
html5lib==1.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.3
multitasking==0.0.11
numpy==1.25.0
pandas==2.0.3
Pillow==10.0.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.17
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
webencodings==0.5.1
Werkzeug==2.3.6
WTForms==3.0.1
yfinance==0.2.22
